11-22

+++ Object_Detection/OD_video_extract.py
@@ -5,6 +5,91 @@ import shutil
 import os
 import glob
 from sortedcontainers import SortedDict
+from CV_text_detection import decode
+
+
+# text detector
+model_path = '/media/weihao/DISK0/Object_detection'
+# model ="C:\\GITHUB\\learnopencv-master\\TextDetectionEAST\\frozen_east_text_detection.pb"
+model = os.path.join(model_path, "frozen_east_text_detection.pb")
+outputLayers = []
+outputLayers.append("feature_fusion/Conv_7/Sigmoid")
+outputLayers.append("feature_fusion/concat_3")
+confThreshold = 0.5  # args.thr
+nmsThreshold = 0.4  # args.nms
+
+
+def text_detection(frame):
+    max_width = 200
+    scale = 1.0
+
+    # Get frame height and width
+    height_ = frame.shape[0]
+    width_ = frame.shape[1]
+
+    if width_ > max_width:
+        scale = max_width / width_
+
+    inpWidth = int(width_ / 32 * scale) * 32
+    inpHeight = int(height_ / 32 * scale) * 32
+
+    rW = width_ / float(inpWidth)
+    rH = height_ / float(inpHeight)
+
+    # Create a 4D blob from frame.
+    blob = cv2.dnn.blobFromImage(frame, 1.0, (inpWidth, inpHeight), (123.68, 116.78, 103.94), True, False)
+
+    # Load network
+    net = cv2.dnn.readNet(model)
+    # Run the model
+    net.setInput(blob)
+    output = net.forward(outputLayers)
+
+    # Get scores and geometry
+    scores = output[0]
+    geometry = output[1]
+    [boxes, confidences] = decode(scores, geometry, confThreshold)
+    t, _ = net.getPerfProfile()
+    label = 'Inference time: %.2f ms' % (t * 1000.0 / cv2.getTickFrequency())
+
+    out_boxes = []
+
+    # Apply NMS
+    indices = cv2.dnn.NMSBoxesRotated(boxes, confidences, confThreshold, nmsThreshold)
+    for i in indices:
+        # get 4 corners of the rotated rect
+        vertices = cv2.boxPoints(boxes[i[0]])
+        # scale the bounding box coordinates based on the respective ratios
+        xmin = 1e6
+        ymin = 1e6
+        xmax = -1
+        ymax = -1
+        for j in range(4):
+            vertices[j][0] *= rW
+            vertices[j][1] *= rH
+            if vertices[j][0] < xmin:
+                xmin = vertices[j][0]
+            if vertices[j][0] > xmax:
+                xmax = vertices[j][0]
+            if vertices[j][1] < ymin:
+                ymin = vertices[j][1]
+            if vertices[j][1] > ymax:
+                ymax = vertices[j][1]
+        xmin = int(xmin) - 10
+        ymin = int(ymin) - 10
+        xmax = int(xmax) + 10
+        ymax = int(ymax) + 10
+        if xmin < 0:
+            xmin = 0
+        if ymin < 0:
+            ymin = 0
+
+        ig = frame[ymin:ymax, xmin:xmax, :]
+        if ig.shape[0] == 0 or ig.shape[1] == 0:
+            continue
+        out_boxes.append(ig)
+    return out_boxes
+
 
 
 def read_p_file(filename):
@@ -40,17 +125,21 @@ def cal_overlap(b1, b2):
 
 
 class Detection:
-    def __init__(self, f, b):
+    def __init__(self, f, id, b, mask):
         self.frame = f
         self.box = b
-        self.persion_id = None
+        self.detection_id = id
+        self.person_id = None
+        self.mask = mask
+        self.text_boxes = None
+
 
 if __name__=='__main__':
     import sys
 
-    filename = sys.argv[1] #'/media/weihao/DISK0/flickr_images/out.mp4'# '/home/weihao/tmp/out.mp4'
-    out_dir = sys.argv[2] #'/home/weihao/tmp/ps'
-
+    filename = '/media/weihao/DISK0/flickr_images/BOLDERBoulder_10K_livestream_at_finish_line.mp4'# '/home/weihao/tmp/out.mp4'
+    out_dir = '/media/weihao/DISK0/flickr_images/ps'
+    cnt = True
 
     persons = []
     nps = {}
@@ -67,12 +156,14 @@ if __name__=='__main__':
             md = dm[a]
             rs = md['rois']
 
-            n1 = 0
-            for p in rs:
-
-                val = Detection(a, p)
+            for n1 in range(len(rs)):
+                if md['class_ids'][n1] != 1:
+                    continue
+                p = rs[n1]
+                mask = md['masks'][n1]
+                val = Detection(a, n1, p, mask)
                 w = p[3]-p[1]
-                if w<50:
+                if w < 50:
                     continue
                 n2 = 0
                 match = False
@@ -88,7 +179,6 @@ if __name__=='__main__':
                 if not match:
                     nps[pid] = [val]
                     pid += 1
-                n1 += 1
 
             i1 = []
             for b in nps:
@@ -105,9 +195,10 @@ if __name__=='__main__':
 
     print("total person", len(persons))
 
-    if os.path.exists(out_dir):
-        shutil.rmtree(out_dir)
-    os.mkdir(out_dir)
+    if not cnt:
+        if os.path.exists(out_dir):
+            shutil.rmtree(out_dir)
+        os.mkdir(out_dir)
 
     rm = []
     for p in persons:
@@ -128,7 +219,8 @@ if __name__=='__main__':
             elif min_w>100:
                 rm.append(p)
             else:
-                print(min_w, max_w)
+                pass
+                # print(min_w, max_w)
 
     n = 0
     for r in rm:
@@ -140,7 +232,8 @@ if __name__=='__main__':
     frame_ids = {}
     pid = 0
     for ps in persons:
-        os.mkdir(out_dir + '/{}'.format(pid))
+        if not cnt:
+            os.mkdir(out_dir + '/{}'.format(pid))
 
         for p in ps:
             if p.frame not in frame_ids:
@@ -148,7 +241,6 @@ if __name__=='__main__':
             p.person_id = pid
             frame_ids[p.frame].append(p)
         pid += 1
-
 
     cap = cv2.VideoCapture(filename)
     nt = 0
@@ -165,15 +257,20 @@ if __name__=='__main__':
                 pid = m.person_id
                 b = m.box
                 ig = clone[b[0]:b[2], b[1]:b[3], :]
-                outname = '{}/{}/I{}.jpg'.format(out_dir, pid, nt)
-                cv2.imwrite(outname, ig)
+                outname = '{}/{}/I{}_{}.jpg'.format(out_dir, pid, nt, m.detection_id)
+                if not os.path.exists(outname):
+                    m.text_boxes = text_detection(ig)
+                    cv2.imwrite(outname, ig)
+                    outname = '{}/{}/I{}_{}.p'.format(out_dir, pid, nt, m.detection_id)
+                    with open(outname, 'wb') as fp:
+                        pickle.dump(m, fp)
                 r = b
                 cv2.line(frame, (r[1], r[0]), (r[3], r[0]), (0, 0, 255), 1, cv2.LINE_AA)
                 cv2.line(frame, (r[3], r[0]), (r[3], r[2]), (0, 0, 255), 1, cv2.LINE_AA)
                 cv2.line(frame, (r[3], r[2]), (r[1], r[2]), (0, 0, 255), 1, cv2.LINE_AA)
                 cv2.line(frame, (r[1], r[2]), (r[1], r[0]), (0, 0, 255), 1, cv2.LINE_AA)
                 cv2.putText(frame, "{}".format(pid), (r[1], r[0]), cv2.FONT_HERSHEY_SIMPLEX, 1, (0,255,0))
-            delay = 16
+            delay = 1
             #
             # if nt not in dm:
             #      break
